requirements.txt
torch
torchvision
transformers
scikit-learn
pillow
tqdm
nltk
rouge-score
numpy 